chore: Update SideNav and Raffles components
The SideNav component was updated to change the navigation link from "/winners" to "/raffles". Additionally, the styling for the active link was modified to remove the background color.

The Raffles component now imports the getRandomWinners function from '../utils/getRandomWinners' and the useNavigate and useLocation hooks from 'react-router-dom'. It also includes a new handleStartDraw function that handles the logic for starting a draw. The handleStartDraw function uploads a CSV file, gets random winners using the getRandomWinners function, and navigates to the '/winners' page with the winners and the uploaded CSV file as state.